User accounts can now be saved to the database (DynamoDBMapper attribute issue). index.html is more responsive. Pushing and moving on to rating/review implementation
Summary: You'll see heavy annotation in the User object. This is necessary for now. Friend id separator in JS is a '-' for now but could be changed to almost anything. I deleted reactapp because it was becoming too difficult and not useful enough for the time required.

Test Plan: User object needs unit tests

Reviewers: zacleids, code-ee

Differential Revision: http://csumbse.net/D570